requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
motor>=3.1.1
requests>=2.28.2
numpy>=1.24.2
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.0.0
pymongo==4.6.1
streamlit>=1.45.0
pandas>=2.0.0
selenium>=4.10.0
webdriver-manager>=3.8.6
charset-normalizer>=3.0.0
psutil>=5.9.0
rich>=12.0.0
plotly>=5.14.0

# Optional testing dependencies
pytest>=7.3.1
pytest-asyncio>=0.21.0


